Animation Update
All units and aliens now have animations that transition accordingly.
When left clicking on an enemy or unit, the health will be displayed.
Adjusted the positions and rotations of the objects that can be picked
up so they will display correctly. Added particle effect to the alien
taint animation, to have it seem as though the alien is spitting a black
gas onto the resources to taint it. When text pops up onto the screen,
it will have a different color and message according to the action that
triggered it.