Added tooltip feature for rocket parts
In the workshop UI, when the user hovers over a part to buy, it will
display the cost in resources for that specific part. Fixed a bug that
was allowing the tooltip panel for objects to be shown even when a menu
was open.  Added floating text for unit abilities when they are
activated and the reactions of the aliens affected by the units ability.
Fixed a bug that was causing units to continue to attack even if they
lost all health the previous frame. Added floating text for when the
game is lost or won. Updated the Game Manager script with a function
that will set all default resource values to 0 and the ship built
boolean to false, so the user will start fresh each time they select new
game. Also, added the credits scene to the build settings, and the
functionality in game to transition to it once the game has been won or
lost.